% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD-package.R
\docType{package}
\name{BDD-package}
\alias{BDD-package}
\alias{_PACKAGE}
\title{BDD: Duplicate Detection Using a Bayesian Partitioning Approach}
\description{
An implementation of the Bayesian duplicate detection model 
    proposed by Sadinle (2014) <DOI:10.1214/14-AOAS779>. It resembles the 
    classic model of Fellegi & Sunter (1969) 
    <DOI:10.1080/01621459.1969.10501049>, but targets a partitioning of the 
    records, rather than pairwise match predictions. Inference is conducted 
    using a Gibbs sampler implemented in C++.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cleanzr/BDD}
  \item Report bugs at \url{https://github.com/cleanzr/BDD/issues}
}

}
\author{
\strong{Maintainer}: Neil Marchant \email{ngmarchant@gmail.com}

Authors:
\itemize{
  \item Rebecca Steorts \email{beka@stat.duke.edu}
  \item Benjamin Rubinstein \email{brubinstein@unimelb.edu.au}
}

}
\keyword{internal}
