% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDDFit.R
\docType{class}
\name{BDDFit-class}
\alias{BDDFit-class}
\title{Fitted Model}
\description{
Stores data associated with a fitted \code{\linkS4class{BDDModel}} object,
including posterior samples, diagnostics, etc.
}
\details{
The \code{history} may include the following \code{\link[coda:mcmc]{coda::mcmc}} objects:
\describe{
\item{links}{A matrix recording samples of the linkage/clustering
structure, where rows correspond to samples and columns correspond to
records. For efficiency reasons, links are only included for
records in the set of candidate pairs. Records not in the set of
candidate pairs are constrained to occupy their own cluster (i.e. they
are prevented from linking with other records). Each sample of the
linkage/clustering structure is encoded as an integer membership
vector: records with the same integer value are assigned to the same
cluster.}
\item{m}{A matrix object recording the m*-probabilities for each
attribute/agreement level. Rows index samples along the Markov chain
and columns index attributes/agreement levels.}
\item{u}{A matrix object recording the u*-probabilities for each
attribute/agreement level. Rows index samples along the Markov chain
and columns index attributes/agreement levels.}
\item{n_clusters}{A vector object recording the number of
clusters/entities the records have been partitioned into. This number
only accounts for records in the set of candidate pairs---i.e. it
excludes singleton clusters associated with records in the set of
non-candidate pairs.}
}
They can be accessed using the \code{\link{extract}} method.
}
\section{Slots}{

\describe{
\item{\code{history}}{A list containing model parameters of interest and summary
statistics along the Markov chain. See below for details.}

\item{\code{state}}{A \code{\linkS4class{BDDModel}} object which represents the state
of the model at the end of the Markov chain.}
}}

