% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDDModel.R
\docType{class}
\name{BDDModel-class}
\alias{BDDModel-class}
\alias{BDDModel}
\title{Container class for the model}
\description{
This class records the state of model parameters that are realized
during MCMC sampling. It also stores indexing data structures for
used to accelerate sampling.
}
\section{Slots}{

\describe{
\item{\code{iteration}}{Non-negative integer. Counts the number of successive
applications of the Markov transition operator. This is not necessarily
equal to the number of samples if thinning is applied.}

\item{\code{m}}{Named list containing the \verb{m*} probabilities for each attribute.
Each entry corresponds to an attribute in \code{comparisons} and contains a
vector of \verb{m*} probabilities for the discrete score levels. Due to the
sequential parameterization of \verb{m*}, there is no probability included
for the largest score level.}

\item{\code{u}}{a named list containing the \verb{u*} probabilities for each attribute.
Each entry corresponds to an attribute in \code{comparisons} and contains a
vector of \verb{u*} probabilities for the discrete score levels. Due to the
sequential parameterization of \verb{u*}, there is no probability included
for the largest score level.}

\item{\code{comparisons}}{a data frame of discrete comparison scores. Each row
corresponds to a pair of records. The discrete comparison scores for
each attribute are stored in columns whose names match the names in
\code{m}, \code{u}, \code{level_counts}, etc.}

\item{\code{candidate_pairs}}{a data frame that specifies pairs of records that
are candidate matches. It contains two columns, \code{ID_1} and \code{ID_2}, which
contain the record identifiers for each pair. Each pair is included
only once: e.g. if pair "A, B" appears, then "B, A" does not appear.}

\item{\code{priors}}{a list containing the priors for \code{alpha1}, \code{alpha0},
\code{beta1}, \code{beta0} and \code{lambda}.}

\item{\code{nc_counts}}{a list containing summary statistics for record pairs that
are non-candidate matches. These are used when updating the \verb{u*}
probabilities.}

\item{\code{row_id_index}}{An index from rec_ids to row indices in \code{candidate_pairs}
where they appear. This is used to speed up inference.}

\item{\code{pair_index}}{An index for \code{candidate_pairs}. Maps a record identifier
to a vector of record identifiers that are potential matches (candidate
pairs).}

\item{\code{level_counts}}{A named list containing counts for each attribute
at each level of agreement for candidate matches.}

\item{\code{rec_ids}}{A named vector that maps the original record identifiers
(the names) to standardized integer identifiers (the entries). Records
that are not candidate matches are excluded.}

\item{\code{links}}{An integer vector of entity identifiers for each record in
\code{rec_ids}. The i-th entry gives the linked entity for the
i-th record in \code{rec_ids}.}
}}

\seealso{
The \code{\link{BDD}} function should be used to initialize a model
from data.
}
