% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{complete_links_samples}
\alias{complete_links_samples}
\title{Extract Complete Samples of the Linkage Structure}
\usage{
complete_links_samples(x, all_rec_ids)
}
\arguments{
\item{x}{An instance of \code{\linkS4class{BDDFit}}. This is the output
of \code{\link{run_inference}}.}

\item{all_rec_ids}{A vector containing the record identifiers for
all records.}
}
\value{
A \code{\link[coda:mcmc]{coda::mcmc}} object where the rows correspond to samples and
the columns correspond to records.
}
\description{
This function extracts samples of the linkage structure which cover \strong{all}
records, including records that are excluded from the candidate pairs.
}
\examples{
## Initialize a BDD model for RLdata500
library(comparator) # provides scoring functions

# Add record ids to the dataframe
RLdata500$ID <- seq.int(nrow(RLdata500))

# Specify model parameters
scoring_fns <- list(
  fname_c1 = Levenshtein(normalize=TRUE),
  lname_c1 = Levenshtein(normalize=TRUE),
  by = BinaryComp(),
  bm = BinaryComp(),
  bd = BinaryComp()
)
scoring_breaks <- list(
  fname_c1 = c(-Inf, .05, .2, .4, Inf),
  lname_c1 = c(-Inf, .05, .2, .4, Inf),
  by = c(-Inf, 0, Inf),
  bm = c(-Inf, 0, Inf),
  bd = c(-Inf, 0, Inf)
)
lambda <- list(
  fname_c1 = c(0.8, 0.85, 0.99),
  lname_c1 = c(0.8, 0.85, 0.99),
  by = c(0.8, 0.99),
  bm = c(0.8, 0.99),
  bd = c(0.8, 0.99)
)

# Prepare pairwise comparison data
pairs <- pairs_all(RLdata500$ID)
pairs <- compute_scores(pairs, RLdata500, scoring_fns, id_col='ID')
pairs <- discretize_scores(pairs, scoring_breaks)
pairs$candidate <- (pairs$fname_c1 < 4) & (pairs$lname_c1 < 4)

model <- BDD(pairs, lambda, candidate_col = 'candidate')

## Run inference
fit <- run_inference(model, n_samples=100, burnin_interval=100)

## Extract samples of the linkage structure
links <- complete_links_samples(fit, RLdata500$ID)

}
\seealso{
In contrast to this function, the \code{\link{extract}} function extracts
samples of the linkage structure for a \emph{subset} of records - only those
records that appear in the candidate pairs.
}
