% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute Pairwise Comparison Scores}
\usage{
compute_scores(pairs, x, scoring_fns, id_col = NULL)
}
\arguments{
\item{pairs}{A data frame or matrix specifying pairs of records in \code{x} to
compare, where rows index pairs and columns index the identifiers of the
two constituent records. If \code{id_col} is specified, the record are
identified according to the corresponding column in \code{x}. Otherwise if
\code{id_col=NULL} the records are identifier by their row index.}

\item{x}{A data frame or matrix of records. Columns must be present for
each attribute referenced in \code{scoring_fns}. An optional column containing
record identifiers is used if \code{id_col} is not NULL. Any other columns
are ignored.}

\item{scoring_fns}{A list containing scoring functions for the attributes
to compare. Each element of the list must be named according to an
attribute in \code{x}. The scoring function must take a pair of vectors as its
first two arguments, and return a vector of scores.}

\item{id_col}{An integer or string specifying a column in \code{x} that contains
record identifiers. Defaults to NULL.}
}
\value{
The \code{pairs} data frame augmented with columns containing scores
for each of the attributes in \code{scoring_fns}.
}
\description{
Given a set of record pairs, this function computes comparison scores
for each attribute. This is known as the "scoring" stage of a record
linkage/entity resolution pipeline.
}
\seealso{
The \code{\link{discretize_scores}} function can be used to discretize the scores
data frame output by this function.
To prepare the record \code{pairs} data frame, refer to the indexing
functions: \code{\link{pairs_all}}, \code{\link{pairs_hamming}} or \code{\link{pairs_fuzzyblock}}.
}
