% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{discretize_scores}
\alias{discretize_scores}
\title{Discretize Pairwise Comparison Scores}
\usage{
discretize_scores(pairs, breaks)
}
\arguments{
\item{pairs}{A data frame of scored record pairs as returned by
\code{\link{compute_scores}}. Must contain the attributes referenced in
\code{breaks}.}

\item{breaks}{A list of scoring breaks for each scores attribute in
\code{pairs}. The names of the entries must match the colnames in \code{pairs}.
The scoring breaks should be represented as a vector of breaks or cut
points, which specify intervals closed on the right.}
}
\value{
A transformed version of the input \code{pairs} data frame, in which
scored columns are discretized. Note that the discrete scores are
encoded as positive integers, where 1 represents the smallest score.
}
\description{
Given a set of scored record pairs, this function converts the scores
to discrete levels.
}
