% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{pairs_all}
\alias{pairs_all}
\title{Index All Record Pairs}
\usage{
pairs_all(rec_ids)
}
\arguments{
\item{rec_ids}{A vector of record identifiers.}
}
\value{
A data frame of record pairs, where rows index pairs and columns
index the identifiers of records in each pair.
}
\description{
Given a collection of \eqn{n} records, this function builds an index
containing all \eqn{n(n-1)/2} record pairs.
}
\seealso{
\code{\link{pairs_hamming}} and \code{\link{pairs_fuzzyblock}} are alternative indexing
functions.
}
