% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{pairs_fuzzyblock}
\alias{pairs_fuzzyblock}
\title{Index Record Pairs Using Fuzzy Blocking}
\usage{
pairs_fuzzyblock(
  x,
  block_col,
  dist_fn,
  dist_cutoff,
  id_col = NULL,
  include_na = FALSE
)
}
\arguments{
\item{x}{A data frame of records. Columns must be present for
each attribute referenced in \code{attribute_cols}. An optional column
containing record identifiers is used if \code{id_col} is not NULL. Any other
columns are ignored.}

\item{block_col}{A string specifying a column in \code{x} to use as the blocking
variable.}

\item{dist_fn}{A vectorized distance function.}

\item{dist_cutoff}{A positive distance cutoff. All record pairs with
a distance less than or equal to this cutoff are included in the index.}

\item{id_col}{A string specifying a column in \code{x} that contains record
identifiers. Defaults to NULL.}

\item{include_na}{A logical specifying whether to include pairs for which
the distance function evaluates to NA. Defaults to FALSE.}
}
\value{
A data frame of record pairs, where rows index pairs and columns
index the identifiers of records in each pair.
}
\description{
Prepares an index of record pairs that are "similar" as determined by a
distance function applied to one of the attributes. This can be viewed as
a fuzzy generalization of standard blocking.
}
\seealso{
\code{\link{pairs_all}} and \code{\link{pairs_hamming}} are alternative indexing
functions
}
