% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{pairs_hamming}
\alias{pairs_hamming}
\title{Index Record Pairs Based on Hamming Distance}
\usage{
pairs_hamming(x, attr_cols, dist_cutoff, id_col = NULL)
}
\arguments{
\item{x}{A data frame of records. Columns must be present for
each attribute referenced in \code{attribute_cols}. An optional column
containing record identifiers is used if \code{id_col} is not NULL. Any other
columns are ignored.}

\item{attr_cols}{A character vector of attribute columns in \code{x} to be
compared when computing the Hamming distance.}

\item{dist_cutoff}{An integer in the interval \verb{[0, length(attribute_cols)]}.
All record pairs with Hamming distance less than or equal to this cutoff
are included in the index.}

\item{id_col}{A string specifying a column in \code{x} that contains record
identifiers. Defaults to NULL.}
}
\value{
A data frame of record pairs, where rows index pairs and columns
index the identifiers of records in each pair.
}
\description{
Given a collection of records, this function builds an index containing
all pairs whose attribute-level Hamming distance falls below some cut-off.
For instance, with a cut-off of 2, all pairs are included that disagree
on 2 or fewer attributes.
}
\seealso{
\code{\link{pairs_all}} and \code{\link{pairs_fuzzyblock}} are alternative
indexing functions
}
